feat: expose legacy_web_url from units in sequence metadata API
we want frontend-app-learning to be able to switch back
and forth between the Course Blocks API and Learning Sequences
API while we transition from the former to the latter.

currently, fa-learning depends on the Course Blocks API
for units' legacy_web_url. The Learning Sequences API is
currnetly not capable of providing that piece of data.

so, we instead expose legacy_web_url from the Sequence Metadata
API. that will free us up to make the Course Blocks API
and Learning Sequences API interchangable in the eyes of
frontend-app-learning.

TNL-8330