fix: Update how the heartbeat routing key is determined.
When the heartbeat was moved to the high priority queue, the production
override did not take into account the fact that the name of the
high_priority queue changes in the production.py file.

Update the part of production.py that sets the default routing key so
that the new HIGH_PRIORITY_QUEUE name is picked up unless there is an
override.