feat: Display an XBlock's version as part of Staff Debug Info
If a course author works on several Open edX instances and they all
have third-party XBlocks installed, it sometimes comes in handy to be
able to tell which version of an XBlock they are working with. For
example, they might want to determine whether a feature that just
recently landed in an XBlock is available to them.

Add an additional field, "block_version", to the context passed to the
"Staff Debug Info" modal. Populate that field by using the inspect
module to sort out the package that the XBlock belongs to, and then
looking up that package's version via the version() method from
importlib.metadata.

In the event that we are unable to discern the package version at
runtime, do not throw a 500. Instead, show block_version as None, and
log a warning.

Related discussion on Open edX Discourse:
https://discuss.openedx.org/t/whats-the-best-way-to-track-show-an-xblocks-version/

(cherry picked from commit 2b32fe5cb563dbfa3675e2e5f41520f8f7ffa9d0)