fix: stop redirecting MFE api requests, and instead give back json
For access denied errors and the like, we were returning 302
redirects to the dashboard or other course pages in our MFE rest
api responses. This would end up causing the MFE to try to parse
html as json and resulted in errors for the user.

This introduces a new 403 response with json content that indicates
where we want to redirect to - the MFE can now read that and
reroute the user as needed.