feat: Allow language preferences to be shared.
We are planning on limiting session cookies for LMS, which
will mean that under the old cookie domain, MFEs and other
IDAs would not have access to the language preference of a user.

This moves language preference cookies to use a different setting,
which means that this cookie can be shared.