feat: add note field to course waffle overrides
It's often very useful to note the reason we override a waffle.
Who asked for it and why - is it temporary until a feature is
completed or a complete opt-out of the feature itself, etc.

Now there's a text field to leave such comments for your future
self, much like waffle flags themselves.