fix: upgrade py2neo from 3.1.2 to 2021.1.5
The dump_to_neo4j management command has not been working
since the upgrade to python 3.8. The latest version of
python that py2neo 3.1.2 states support for is python 3.5,
so this isn't surprising.

The earliest non-prerelease version of py2neo that supports
python 3.8 is 2020.x (skipping the 4.x and 5.x series). Since
we're going as far as a 2020.x, we may as well upgrade all the
way to the newest series, 2021.x.

This will require an upgrade of Coursegraph's Neo4j
version from 3.2.x to 3.5.x.

TNL-8386