feat: make sequence metadata API handle and provide hashed usage keys
We want to enable shorter URLs in the courseware pages of
frontend-app-learning.

To do this, we define a function hash_usage_key,
which maps usage keys to an 8-character-
long, urlsafe-base64-encoded hash. The function is
defined in the learning_sequences data model, and hashes
of sequence keys are persisted to learning_sequences upon
outline generation. We also present an experimental
learning_sequences python API for fetching sequences by
their hash (instead of their usage key).

Building on this, we enhance the Sequence Metadata
HTTP API to accept hashes as sequence identifiers
(in addition to usage keys). The API also now returns
the usage_key_hash of the requested sequence, as
well as the usage_key_hashes of every unit it returns
as part of the sequence. We may revisit the exposure
of usage_key_hashes from this API in
https://openedx.atlassian.net/browse/TNL-8638,
as we think it may be better to only expose this data
from the learning_sequences Outlines HTTP API.

Finally, we should note that this removes one
constraint from the LearningSequence model:
    <unique together on (learning_context_key, usage_key)>
and adds two new ones:
    <unique on usage_key>
    <unique together on (learning_context_key, usage_key_hash)>
in order to support the new learning_sequences

Public Jira (archived): https://openedx.atlassian.net/browse/TNL-8511

Private-ref: https://2u-internal.atlassian.net/browse/TNL-8511
Co-Authored-By: Kristin Aoki <kaoki@edx.org>