fix: Google Analytics tracking in LMS
Google Analytics tracking was broken in the LMS because the `escapejs` function
was replacing the dash "-" by "\u002D". As far as we understand, this issue was
introduced in 2017 by the following change:
https://github.com/edx/edx-platform/commit/7ea3b013ad47eab5b6532e3af602892894781d55#diff-1adbd21958e452c19656801dc282c38d458be2b2afeecad134e77c1de607a74dR117

The solution consists of not escaping the GOOGLE_ANALYTICS_ACCOUNT setting.
Generally speaking, settings and site configuration should be sufficiently
trusted that we don't need to escape them.

This issue was first reported here: https://discuss.overhang.io/t/utf16-encoding-of-plugin-string-breaking-google-analytics/1756/