temp: Use a custom schema class for pulsar messages.
Instead of using the built-in record classes, use a custom schema class
that let's us have better controll of the avro schema by defining it
explicitly rather than implicitly via Record and Field declerations.

Using `Record` classes was not super useful because it was not possible
to create Optional or Required fields that had default values other than
None.