fix: amend type annotation that was strengthened by xmodule extraction
There was an annotation in enrollment_track_partition_groups.py
that previously was accepted by mypy, but began failing once xmodule
was elevated to a top-level package. My assumption is that this is
because mypy is now able to resolve `xmodule` files as local modules,
thus making type-checking for xmodule-related code slightly stricter.

Thus, this commit tweaks a type annotation to be more correct: it adds
an explicit cast on a dynamically-typed value being returned by
get_user_partition_groups.