fix(program-portals): prevent resume course links from 404ing (#27046)
* fix(program-portals): prevent resume course links from 404ing

From time to time course teams will delete a problem in an active
course. When this happens and that particular problem is the last
completed subsection for a learner, some naive attempts to provide a
"continue course" that takes the learner back to where they left off
may result in 404s. This checks whether a learner would be able to
access the block returned from completion data before we try to create
a link to that part of the course structure to hand to that learner.
It's used on the Master's program learner portal pages.

JIRA:EDUCATOR-5313

Co-authored-by: Zach Hancock <zhancock@edx.org>