fix: fix site-wide language for MFEs
When a site-wide language is set in site configurations, the desired
result is that MFEs use that language over the one which is set in user
preferences (or if None is set, it would fallback to 'en').

Since the MFEs use language value from cookies to determine which
language to use, and the existing solution was only modifying the
language in the session, and not overwrite the value in the cookies, the
MFEs would use language set in users' preferences (or 'en' if nothing
was set).

This commit removes the existing solution and implements a new one in
the middleware that sets language values in cookies based on the values
set in users' preferences. The modified middleware is going to check if
there is a site-wide language set in site configurations, and if it is,
will use that value over the one which is set in users' preferences.

Additionally, if site-wide language is set, the middleware won't save
the language which is passed in cookies as the user preference, to not
overwrite the user preference for other sites, where this site
configuration doesn't apply.