fix: consider course end date in visible_date
We send a visible_date from Platform to Credentials so that credentials
knows what date to display on the learner record (also used in some
other places, like calculating program cert date). If the course is
instructor-paced and the certificate display behavior is set to "end",
we need to send the course end date as the visible date.

Related to MICROBA-1554.