feat: ESG BFF submission/lock endpoint (#29270)
* feat: add util for unpacking query params

require_params loads the provided params to the view function and will
return a 400 if the param is not provided. This should cut down on
boilerplate.

* refactor: use require_params to remove boilerplate

* refactor: add read_only to serializers

* feat: extend call_xblock_json_handler

Add decode kwarg. Setting False (non-default) returns the entire
HttpResponse instead of unpacked JSON body. Helpful if you want to
preserve the status code in your request or inspect other response info

* feat: add lock status serializer.

* feat: add submission lock views

* feat: add submission lock path to URLs

* refactor: return error codes instead of long-form