feat!: improve discussion provider listing and configuration
Includes the new discussion provider in the API response under certain circumstance. If the
legacy provider is already in use, the new provider will be hidden. If the new provider is used
then the legacy provider will be hidden. If some other provider is used, then the new or legacy
providers will be visible depending on whether the discussions MFE is enabled.

This also fixes the "group_at_subsection" setting so that it's saved when using the new provider,
and also removes legacy plugin configuraiton when the legacy or new provider aren't being used.